feat(inventory): implement stock control management and inventory tracking
- Implement `InventoryBloc`, `InventoryUseCase`, and `InventoryRepoImpl` to handle stock data fetching, pagination, and filtering.
- Create `InventoryScreen` featuring a stock summary dashboard (Total, Low Stock, Out of Stock) and a paginated stock list.
- Add `StockResponseDto` and `StockData` models to support inventory-related API responses.
- Integrate `InventoryScreen` into the main navigation hub and update `FeatureMenuSections` with a "Stock Control" entry.
- Update `ProductData` and `StockChip` to handle stock levels as `double` instead of `int` for better precision.
- Register inventory dependencies in `dependencies_provider.dart` and `providers.dart`.
- Refactor `FeatureMenu` by removing the redundant `UserHeader` widget to streamline the menu structure.
- Clean up `NavigationConfig` for better code formatting and readability.

+++ lib/features/products/presentation/widgets/stock_chip.dart
@@ -2,7 +2,7 @@ import 'package:amana_pos/theme/app_text_styles.dart';
 import 'package:flutter/material.dart';
 
 class StockChip extends StatelessWidget {
-  final int level;
+  final double level;
   const StockChip({super.key, required this.level});
 
   @override
